Misc improvements to the `@resumable` macro (#71)
- more descriptive names generated symbols
- enable docstrings for `@resumable` functions
- do not delete line number nodes in `@resumable` functions

Fixes #39

Fixes #45

Initial steps needed to address #46

* better names for generated symbols

* make docstrings work with @resumable

* do not throw away line number information

* tests for line number and docstring preservation

* readme cleanup and version bump

* relax JET tests

+++ benchmark/benchmarks.jl
@@ -3,7 +3,7 @@ using ResumableFunctions
 
 const n = 93
 
-const N = BigInt
+const N = Int
 
 function direct(a::N, b::N)
   b, a+b
@@ -121,8 +121,8 @@ end
 @noinline function test_closure_opt(n::Int)
   fib_closure = fibonacci_closure_opt()
   a = 0
-  for _ in 1:n 
-    a = fib_closure() 
+  for _ in 1:n
+    a = fib_closure()
   end
   a
 end
